sjogreen_x is implemented and run in double precision. seems correct

+++ example/wallshock_x/main.h
@@ -70,7 +70,7 @@ int main(int narg, char **args){
   /*
     initalize the positions of particles
   */
-  for(int i=0; i<n_cell; ++i){
+  for(size_t i=0; i<n_cell; ++i){
     alloc_uniform_random(
       make_cell_at(c, i, 0, 0),
       n_particle_per_cell,
